backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
aiosqlite==0.20.0
greenlet==3.0.3
python-dateutil==2.9.0
reportlab==4.2.0
fpdf2==2.7.9
openpyxl==3.1.2
Pillow==10.3.0
pandas==2.2.2
httpx==0.27.0

# Standalone desktop packaging
pystray==0.19.5
pyinstaller==6.6.0
